requirements-dev.txt
selenium
coverage
fabric3